Time: 4 ms (90.91%) | Memory: 203.7 MB (95.45%) - LeetSync

+++ 2126-count-nodes-equal-to-sum-of-descendants/count-nodes-equal-to-sum-of-descendants.cpp
@@ -11,7 +11,7 @@
  */
 class Solution {
 public:
-    long long count=0;
+    int count=0;
     long long helper(TreeNode* root,long long sum){
         if(root==NULL){
             return 0;
